requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.2.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.3
django-bootstrap-v5==1.0.11
django-import-export==3.2.0
et-xmlfile==1.1.0
idna==3.4
MarkupPy==1.14
numpy==1.25.1
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.0.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tablib==3.5.0
tzdata==2023.3
urllib3==2.0.3
xlrd==2.0.1
xlwt==1.3.0
